src/main/java/ru/netology/Game.java
package ru.netology;

import java.util.ArrayList;

public class Game {

    private ArrayList<Player> players = new ArrayList<>();

    public Player register(Player player) {
        players.add(player);
            return player;
    }

    public Player findByName(String name) {
        for (Player player : players) {
            if (player.getName().equals(name)) {
                return player;
            }
        }
        return null;
    }

    public int round(String playerName1, String playerName2) {
        if (findByName(playerName1) == null) {
            throw new NotRegisteredException("Player " + playerName1 + " is not registered");
        } else if (findByName(playerName2) == null) {
            throw new NotRegisteredException("Player" + playerName2 + " is not registered");
        }

        if (findByName(playerName1).getStrength() > findByName(playerName2).getStrength()) {
            return 1;
        } else if (findByName(playerName1).getStrength() < findByName(playerName2).getStrength()) {
            return -1;
        } else {
            return 0;
        }
    }
}
